Start working on the preliminary bill of materials
All funding needs to be preapproved, even though the funds are distributed through reimbursement. This means even though nothing has been purchased, and I will have to purchase everything on my own, I have to provide a bill of materials with almost exact pricing up front. This makes it extremely difficult to pivot or change design down the line, so I've got to do a bunch of research up front to figure out what's going on