Finish initial draft of pcb BOM
Because I don't have much experience with PCB making, I've been trying to estimate what I'll need for this project to work so I can submit the funding application for pre-approval. Since everything has to be pre-approved, I need to have a general idea of what I want even though I don't have the experience needed to know what I want, and to see a couple steps down the line for potential issues.

So what I've done is compiled a potential list, and once funding is approved I will order the basic materials like the io board so I can get my hands on the cm4 and then redesign probably redesign everything then.

Note this isn't the final BOM that will be submitted because correct quantities are not filled out yet. That will be next.