Funding Approved!
Huzzah! The funding application has been application has been approved! I've download a copy of the receipt for records and will merge this in now that it is fully resolved. I think I'll keep record of all the receipts in a separate folder, but I'll keep uploading them as I purchase items.