Initialize code directory
Create folders for the forked moonlight code, the arm image profile data, and scripts to automate everything. Everything is subject to change since technically a lot of this can only be run on a manjaro system, so I'm going to see how I want to handle everything. All of this is going to be new for me so there will be a lot of experimentation.