Record first points of data
All of this is subject to change as I don't really know the best way to do this. What I've record so far is the input delay for the moonlight application determined by the delay in keystrokes and color changes as recorded by the code in the `datalogger` folder. Right now these are just csv files that can be read by latex, but I might look for a more organized solution soon when I record data for the other platforms