Prepare Complete (not final) draft
This draft will be sent of to each of the committee members to read through before the thesis defense.
The PDF has been rendered in "final" mode, so no comments will show through.

small.pdf is a compressed version of the main pdf, with no differences except each of the images have been compressed to 300dpi.